backend/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.7
flask-cors==4.0.0
requests==2.31.0
